More style/consistency improvements

+++ src/Canvas.hpp
@@ -13,7 +13,7 @@ class Canvas {
   using ColorRow = std::vector<Color>;
   using ColorMatrix = std::vector<ColorRow>;
 
-  Canvas(ptrdiff_t width, ptrdiff_t height) noexcept
+  Canvas(int width, int height) noexcept
       : pixels_(static_cast<ColorMatrix::size_type>(height),
                 ColorRow(static_cast<ColorRow::size_type>(width),
                          Color(0.f, 0.f, 0.f))) {
@@ -22,26 +22,26 @@ class Canvas {
 
   [[nodiscard]] ColorMatrix pixels() const noexcept { return pixels_; }
 
-  [[nodiscard]] ptrdiff_t width() const noexcept {
+  [[nodiscard]] int width() const noexcept {
     if (pixels_.empty()) return 0;
 
-    return static_cast<ptrdiff_t>(pixels_[0].size());
+    return static_cast<int>(pixels_[0].size());
   }
 
-  [[nodiscard]] ptrdiff_t height() const noexcept {
-    return static_cast<ptrdiff_t>(pixels_.size());
+  [[nodiscard]] int height() const noexcept {
+    return static_cast<int>(pixels_.size());
   }
 
   [[nodiscard]] bool empty() const noexcept { return pixels_.empty(); }
 
-  void write_pixel(ptrdiff_t x, ptrdiff_t y, const Color& color) noexcept {
+  void write_pixel(int x, int y, const Color& color) noexcept {
     assert(x < width() && x >= 0 && y < height() && y >= 0);
 
     pixels_[static_cast<ColorMatrix::size_type>(y)]
            [static_cast<ColorRow::size_type>(x)] = color;
   }
 
-  [[nodiscard]] Color pixel_at(ptrdiff_t x, ptrdiff_t y) const noexcept {
+  [[nodiscard]] Color pixel_at(int x, int y) const noexcept {
     return pixels_[static_cast<ColorMatrix::size_type>(y)]
                   [static_cast<ColorRow::size_type>(x)];
   }
@@ -50,8 +50,7 @@ class Canvas {
   ColorMatrix pixels_{};
 };
 
-[[nodiscard]] bool in_range(const Canvas& c, ptrdiff_t x,
-                            ptrdiff_t y) noexcept {
+[[nodiscard]] bool in_range(const Canvas& c, int x, int y) noexcept {
   return x < c.width() && x >= 0 && y < c.height() && y >= 0;
 }
